requirements.txt
Django==1.9.9
django-geojson==2.9.0
django-leaflet==0.18.2
pkg-resources==0.0.0
psycopg2==2.6.2
six==1.10.0
